Upgrade tests
* Pytest functions, no classes or unittest functions required
* One file per part of the public API

Test Plan: Checked the numbers of passed/skipped/xfailed tests on both Python 2 and 3 match before/after.
